@@ -52,41 +60,43 @@
     }
   ],
   "dependencies": {
-    "mime-types": "~2.1.16",
+    "mime-types": "~2.1.18",
     "negotiator": "0.6.1"
   },
   "description": "Higher-level content negotiation",
   "devDependencies": {
-    "eslint": "3.19.0",
-    "eslint-config-standard": "10.2.1",
-    "eslint-plugin-import": "2.7.0",
+    "eslint": "4.18.1",
+    "eslint-config-standard": "11.0.0",
+    "eslint-plugin-import": "2.9.0",
     "eslint-plugin-markdown": "1.0.0-beta.6",
-    "eslint-plugin-node": "5.1.1",
-    "eslint-plugin-promise": "3.5.0",
+    "eslint-plugin-node": "6.0.1",
+    "eslint-plugin-promise": "3.6.0",
     "eslint-plugin-standard": "3.0.1",
     "istanbul": "0.4.5",
     "mocha": "~1.21.5"
   },
   "directories": {},
   "dist": {
-    "shasum": "86246758c7dd6d21a6474ff084a4740ec05eb21f",
-    "tarball": "https://registry.npmjs.org/accepts/-/accepts-1.3.4.tgz"
+    "shasum": "eb777df6011723a3b14e8a72c0805c8e86746bd2",
+    "tarball": "https://registry.npmjs.org/accepts/-/accepts-1.3.5.tgz",
+    "fileCount": 5,
+    "unpackedSize": 16555
   },
   "engines": {
     "node": ">= 0.6"
   },
   "files": [
-    "HISTORY.md",
     "LICENSE",
+    "HISTORY.md",
     "index.js"
   ],
-  "gitHead": "71ea430741d6eb5484b6c67c95924540a98186a5",
+  "gitHead": "c38d0e968cdc1526f7cc7a718977ee76655c84f5",
   "homepage": "https://github.com/jshttp/accepts#readme",
   "keywords": [
-    "accept",
-    "accepts",
     "content",
-    "negotiation"
+    "negotiation",
+    "accept",
+    "accepts"
   ],
   "license": "MIT",
   "maintainers": [
