@@ -44,34 +61,35 @@
   "dependencies": {},
   "description": "Convert a dash/dot/underscore/space separated string to camelCase: foo-bar → fooBar",
   "devDependencies": {
-    "ava": "0.0.4"
+    "ava": "*",
+    "xo": "*"
   },
   "directories": {},
   "dist": {
-    "shasum": "9bb5304d2e0b56698b2c758b08a3eaa9daa58a39",
-    "tarball": "https://registry.npmjs.org/camelcase/-/camelcase-1.2.1.tgz"
+    "shasum": "d545635be1e33c542649c69173e5de6acfae34dd",
+    "tarball": "https://registry.npmjs.org/camelcase/-/camelcase-4.1.0.tgz"
   },
   "engines": {
-    "node": ">=0.10.0"
+    "node": ">=4"
   },
   "files": [
     "index.js"
   ],
-  "gitHead": "185ba12da723be9c1ee986cc2956bdc4c517a141",
-  "homepage": "https://github.com/sindresorhus/camelcase",
+  "gitHead": "0e6e4a2752aa013b8e9477145c7b8132c95a82ef",
+  "homepage": "https://github.com/sindresorhus/camelcase#readme",
   "keywords": [
-    "camel",
-    "camel-case",
     "camelcase",
+    "camel-case",
+    "camel",
     "case",
-    "convert",
     "dash",
-    "dot",
     "hyphen",
+    "dot",
+    "underscore",
     "separator",
     "string",
     "text",
-    "underscore"
+    "convert"
   ],
   "license": "MIT",
   "maintainers": [
