@@ -1,87 +1,98 @@
 {
   "_args": [
     [
-      "finalhandler@~1.0.6",
-      "/home/gil/Application/nodejs/sbms/node_modules/express"
+      {
+        "raw": "finalhandler@1.1.1",
+        "scope": null,
+        "escapedName": "finalhandler",
+        "name": "finalhandler",
+        "rawSpec": "1.1.1",
+        "spec": "1.1.1",
+        "type": "version"
+      },
+      "C:\\Android_react\\PPSI-11\\node_modules\\express"
     ]
   ],
-  "_from": "finalhandler@>=1.0.6 <1.1.0",
-  "_id": "finalhandler@1.0.6",
+  "_from": "finalhandler@1.1.1",
+  "_id": "finalhandler@1.1.1",
   "_inCache": true,
-  "_installable": true,
   "_location": "/finalhandler",
-  "_nodeVersion": "6.11.1",
+  "_nodeVersion": "6.13.0",
   "_npmOperationalInternal": {
     "host": "s3://npm-registry-packages",
-    "tmp": "tmp/finalhandler-1.0.6.tgz_1506104827410_0.847819667076692"
+    "tmp": "tmp/finalhandler_1.1.1_1520357803698_0.8904166922156158"
   },
   "_npmUser": {
-    "email": "doug@somethingdoug.com",
-    "name": "dougwilson"
+    "name": "dougwilson",
+    "email": "doug@somethingdoug.com"
   },
-  "_npmVersion": "3.10.10",
+  "_npmVersion": "5.6.0",
   "_phantomChildren": {},
   "_requested": {
-    "name": "finalhandler",
-    "raw": "finalhandler@~1.0.6",
-    "rawSpec": "~1.0.6",
+    "raw": "finalhandler@1.1.1",
     "scope": null,
-    "spec": ">=1.0.6 <1.1.0",
-    "type": "range"
+    "escapedName": "finalhandler",
+    "name": "finalhandler",
+    "rawSpec": "1.1.1",
+    "spec": "1.1.1",
+    "type": "version"
   },
   "_requiredBy": [
     "/express"
   ],
-  "_resolved": "https://registry.npmjs.org/finalhandler/-/finalhandler-1.0.6.tgz",
-  "_shasum": "007aea33d1a4d3e42017f624848ad58d212f814f",
+  "_resolved": "https://registry.npmjs.org/finalhandler/-/finalhandler-1.1.1.tgz",
+  "_shasum": "eebf4ed840079c83f4249038c9d703008301b105",
   "_shrinkwrap": null,
-  "_spec": "finalhandler@~1.0.6",
-  "_where": "/home/gil/Application/nodejs/sbms/node_modules/express",
+  "_spec": "finalhandler@1.1.1",
+  "_where": "C:\\Android_react\\PPSI-11\\node_modules\\express",
   "author": {
-    "email": "doug@somethingdoug.com",
-    "name": "Douglas Christopher Wilson"
+    "name": "Douglas Christopher Wilson",
+    "email": "doug@somethingdoug.com"
   },
   "bugs": {
     "url": "https://github.com/pillarjs/finalhandler/issues"
   },
   "dependencies": {
     "debug": "2.6.9",
-    "encodeurl": "~1.0.1",
+    "encodeurl": "~1.0.2",
     "escape-html": "~1.0.3",
     "on-finished": "~2.3.0",
     "parseurl": "~1.3.2",
-    "statuses": "~1.3.1",
+    "statuses": "~1.4.0",
     "unpipe": "~1.0.0"
   },
   "description": "Node.js final http responder",
   "devDependencies": {
-    "eslint": "3.19.0",
-    "eslint-config-standard": "10.2.1",
-    "eslint-plugin-import": "2.7.0",
+    "eslint": "4.18.1",
+    "eslint-config-standard": "11.0.0",
+    "eslint-plugin-import": "2.9.0",
     "eslint-plugin-markdown": "1.0.0-beta.6",
-    "eslint-plugin-node": "5.1.1",
-    "eslint-plugin-promise": "3.5.0",
+    "eslint-plugin-node": "6.0.1",
+    "eslint-plugin-promise": "3.6.0",
     "eslint-plugin-standard": "3.0.1",
     "istanbul": "0.4.5",
     "mocha": "2.5.3",
-    "readable-stream": "2.3.3",
+    "readable-stream": "2.3.4",
     "safe-buffer": "5.1.1",
     "supertest": "1.1.0"
   },
   "directories": {},
   "dist": {
-    "shasum": "007aea33d1a4d3e42017f624848ad58d212f814f",
-    "tarball": "https://registry.npmjs.org/finalhandler/-/finalhandler-1.0.6.tgz"
+    "integrity": "sha512-Y1GUDo39ez4aHAw7MysnUD5JzYX+WaIj8I57kO3aEPT1fFRL4sr7mjei97FgnwhAyyzRYmQZaTHb2+9uZ1dPtg==",
+    "shasum": "eebf4ed840079c83f4249038c9d703008301b105",
+    "tarball": "https://registry.npmjs.org/finalhandler/-/finalhandler-1.1.1.tgz",
+    "fileCount": 5,
+    "unpackedSize": 16905
   },
   "engines": {
     "node": ">= 0.8"
   },
   "files": [
-    "HISTORY.md",
     "LICENSE",
+    "HISTORY.md",
     "index.js"
   ],
-  "gitHead": "ed4c24d4d7f78b3136ca7d0e7215541cb921a980",
+  "gitHead": "024f493418f62a59592a98f07b23b265092c1006",
   "homepage": "https://github.com/pillarjs/finalhandler#readme",
   "license": "MIT",
   "maintainers": [
