@@ -1,75 +1,93 @@
 {
   "_args": [
     [
-      "proxy-addr@~1.1.5",
-      "/home/gil/Application/nodejs/sbms/node_modules/express"
+      {
+        "raw": "proxy-addr@~2.0.3",
+        "scope": null,
+        "escapedName": "proxy-addr",
+        "name": "proxy-addr",
+        "rawSpec": "~2.0.3",
+        "spec": ">=2.0.3 <2.1.0",
+        "type": "range"
+      },
+      "C:\\Android_react\\PPSI-11\\node_modules\\express"
     ]
   ],
-  "_from": "proxy-addr@>=1.1.5 <1.2.0",
-  "_id": "proxy-addr@1.1.5",
+  "_from": "proxy-addr@>=2.0.3 <2.1.0",
+  "_id": "proxy-addr@2.0.3",
   "_inCache": true,
-  "_installable": true,
   "_location": "/proxy-addr",
-  "_nodeVersion": "6.11.1",
+  "_nodeVersion": "6.13.0",
   "_npmOperationalInternal": {
     "host": "s3://npm-registry-packages",
-    "tmp": "tmp/proxy-addr-1.1.5.tgz_1501030838206_0.9159589342307299"
+    "tmp": "tmp/proxy-addr_2.0.3_1519101976942_0.28350766039209363"
   },
   "_npmUser": {
-    "email": "doug@somethingdoug.com",
-    "name": "dougwilson"
+    "name": "dougwilson",
+    "email": "doug@somethingdoug.com"
   },
-  "_npmVersion": "3.10.10",
+  "_npmVersion": "5.6.0",
   "_phantomChildren": {},
   "_requested": {
-    "name": "proxy-addr",
-    "raw": "proxy-addr@~1.1.5",
-    "rawSpec": "~1.1.5",
+    "raw": "proxy-addr@~2.0.3",
     "scope": null,
-    "spec": ">=1.1.5 <1.2.0",
+    "escapedName": "proxy-addr",
+    "name": "proxy-addr",
+    "rawSpec": "~2.0.3",
+    "spec": ">=2.0.3 <2.1.0",
     "type": "range"
   },
   "_requiredBy": [
     "/express"
   ],
-  "_resolved": "https://registry.npmjs.org/proxy-addr/-/proxy-addr-1.1.5.tgz",
-  "_shasum": "71c0ee3b102de3f202f3b64f608d173fcba1a918",
+  "_resolved": "https://registry.npmjs.org/proxy-addr/-/proxy-addr-2.0.3.tgz",
+  "_shasum": "355f262505a621646b3130a728eb647e22055341",
   "_shrinkwrap": null,
-  "_spec": "proxy-addr@~1.1.5",
-  "_where": "/home/gil/Application/nodejs/sbms/node_modules/express",
+  "_spec": "proxy-addr@~2.0.3",
+  "_where": "C:\\Android_react\\PPSI-11\\node_modules\\express",
   "author": {
-    "email": "doug@somethingdoug.com",
-    "name": "Douglas Christopher Wilson"
+    "name": "Douglas Christopher Wilson",
+    "email": "doug@somethingdoug.com"
   },
   "bugs": {
     "url": "https://github.com/jshttp/proxy-addr/issues"
   },
   "dependencies": {
-    "forwarded": "~0.1.0",
-    "ipaddr.js": "1.4.0"
+    "forwarded": "~0.1.2",
+    "ipaddr.js": "1.6.0"
   },
   "description": "Determine address of proxied request",
   "devDependencies": {
     "beautify-benchmark": "0.2.4",
     "benchmark": "2.1.4",
-    "istanbul": "0.4.5",
-    "mocha": "~1.21.5"
+    "eslint": "4.18.0",
+    "eslint-config-standard": "11.0.0",
+    "eslint-plugin-import": "2.8.0",
+    "eslint-plugin-markdown": "1.0.0-beta.6",
+    "eslint-plugin-node": "6.0.0",
+    "eslint-plugin-promise": "3.6.0",
+    "eslint-plugin-standard": "3.0.1",
+    "mocha": "3.5.3",
+    "nyc": "10.3.2"
   },
   "directories": {},
   "dist": {
-    "shasum": "71c0ee3b102de3f202f3b64f608d173fcba1a918",
-    "tarball": "https://registry.npmjs.org/proxy-addr/-/proxy-addr-1.1.5.tgz"
+    "integrity": "sha512-jQTChiCJteusULxjBp8+jftSQE5Obdl3k4cnmLA6WXtK6XFuWRnvVL7aCiBqaLPM8c4ph0S4tKna8XvmIwEnXQ==",
+    "shasum": "355f262505a621646b3130a728eb647e22055341",
+    "tarball": "https://registry.npmjs.org/proxy-addr/-/proxy-addr-2.0.3.tgz",
+    "fileCount": 5,
+    "unpackedSize": 15421
   },
   "engines": {
-    "node": ">= 0.6"
+    "node": ">= 0.10"
   },
   "files": [
-    "HISTORY.md",
     "LICENSE",
+    "HISTORY.md",
     "README.md",
     "index.js"
   ],
-  "gitHead": "f40ceab074ec2f92399d112793d9ad1c9d96e146",
+  "gitHead": "d841ab7a2150963793ad2dc121df014f426f2964",
   "homepage": "https://github.com/jshttp/proxy-addr#readme",
   "keywords": [
     "ip",
