@@ -53,35 +62,40 @@
   "dependencies": {},
   "description": "HTTP status utility",
   "devDependencies": {
-    "csv-parse": "1.1.7",
-    "eslint": "3.10.0",
-    "eslint-config-standard": "6.2.1",
-    "eslint-plugin-promise": "3.3.2",
-    "eslint-plugin-standard": "2.0.1",
+    "csv-parse": "1.2.4",
+    "eslint": "3.19.0",
+    "eslint-config-standard": "10.2.1",
+    "eslint-plugin-import": "2.8.0",
+    "eslint-plugin-markdown": "1.0.0-beta.6",
+    "eslint-plugin-node": "5.2.0",
+    "eslint-plugin-promise": "3.6.0",
+    "eslint-plugin-standard": "3.0.1",
     "istanbul": "0.4.5",
     "mocha": "1.21.5",
+    "raw-body": "2.3.2",
     "stream-to-array": "2.3.0"
   },
   "directories": {},
   "dist": {
-    "shasum": "faf51b9eb74aaef3b3acf4ad5f61abf24cb7b93e",
-    "tarball": "https://registry.npmjs.org/statuses/-/statuses-1.3.1.tgz"
+    "integrity": "sha512-zhSCtt8v2NDrRlPQpCNtw/heZLtfUDqxBM1udqikb/Hbk52LK4nQSwr10u77iopCW5LsyHpuXS0GnEc48mLeew==",
+    "shasum": "bb73d446da2796106efcc1b601a253d6c46bd087",
+    "tarball": "https://registry.npmjs.org/statuses/-/statuses-1.4.0.tgz"
   },
   "engines": {
     "node": ">= 0.6"
   },
   "files": [
     "HISTORY.md",
-    "LICENSE",
+    "index.js",
     "codes.json",
-    "index.js"
+    "LICENSE"
   ],
-  "gitHead": "28a619be77f5b4741e6578a5764c5b06ec6d4aea",
-  "homepage": "https://github.com/jshttp/statuses",
+  "gitHead": "f76682144d9f0ed2c726bf0a8c868a33e393a8e5",
+  "homepage": "https://github.com/jshttp/statuses#readme",
   "keywords": [
-    "code",
     "http",
-    "status"
+    "status",
+    "code"
   ],
   "license": "MIT",
   "maintainers": [
