@@ -171,36 +184,60 @@
     {
       "name": "Gilad Peleg",
       "email": "giladp007@gmail.com"
+    },
+    {
+      "name": "djchie",
+      "email": "djchie.dev@gmail.com"
+    },
+    {
+      "name": "Gary Ye",
+      "email": "garysye@gmail.com"
+    },
+    {
+      "name": "Nicolas Lalevée",
+      "email": "nicolas.lalevee@hibnet.org"
     }
   ],
-  "dependencies": {
-    "amdefine": ">=0.0.4"
-  },
+  "dependencies": {},
   "description": "Generates and consumes source maps",
   "devDependencies": {
-    "dryice": ">=0.4.8"
+    "doctoc": "^0.15.0",
+    "webpack": "^1.12.0"
   },
-  "directories": {
-    "lib": "./lib"
-  },
+  "directories": {},
   "dist": {
-    "shasum": "eba4f5da9c0dc999de68032d8b4f76173652036b",
-    "tarball": "https://registry.npmjs.org/source-map/-/source-map-0.4.4.tgz"
+    "integrity": "sha512-UjgapumWlbMhkBgzT7Ykc5YXUT46F0iKu8SGXq0bcwP5dz/h0Plj6enJqjz1Zbq2l5WaqYnrVbwWOWMyF3F47g==",
+    "shasum": "74722af32e9614e9c287a8d0bbde48b5e2f1a263",
+    "tarball": "https://registry.npmjs.org/source-map/-/source-map-0.6.1.tgz"
   },
   "engines": {
-    "node": ">=0.8.0"
+    "node": ">=0.10.0"
   },
   "files": [
-    "build/",
-    "lib/"
+    "source-map.js",
+    "source-map.d.ts",
+    "lib/",
+    "dist/source-map.debug.js",
+    "dist/source-map.js",
+    "dist/source-map.min.js",
+    "dist/source-map.min.js.map"
   ],
+  "gitHead": "ac518d2f21818146f3310557bd51c13d8cff2ba8",
   "homepage": "https://github.com/mozilla/source-map",
   "license": "BSD-3-Clause",
-  "main": "./lib/source-map.js",
+  "main": "./source-map.js",
   "maintainers": [
     {
+      "name": "tromey",
+      "email": "tom@tromey.com"
+    },
+    {
+      "name": "ejpbruel",
+      "email": "ejpbruel@gmail.com"
+    },
+    {
       "name": "mozilla-devtools",
-      "email": "mozilla-developer-tools@googlegroups.com"
+      "email": "nfitzgerald@mozilla.com"
     },
     {
       "name": "mozilla",
