@@ -1,145 +1,146 @@
 {
   "_args": [
     [
-      "yargs@~3.10.0",
-      "/home/gil/Application/nodejs/sbms/node_modules/uglify-js"
+      {
+        "raw": "yargs@^8.0.2",
+        "scope": null,
+        "escapedName": "yargs",
+        "name": "yargs",
+        "rawSpec": "^8.0.2",
+        "spec": ">=8.0.2 <9.0.0",
+        "type": "range"
+      },
+      "C:\\Android_react\\PPSI-11\\node_modules\\react-native-vector-icons"
     ]
   ],
-  "_from": "yargs@>=3.10.0 <3.11.0",
-  "_id": "yargs@3.10.0",
+  "_from": "yargs@>=8.0.2 <9.0.0",
+  "_id": "yargs@8.0.2",
   "_inCache": true,
-  "_installable": true,
   "_location": "/yargs",
-  "_nodeVersion": "2.0.2",
-  "_npmUser": {
-    "email": "ben@npmjs.com",
-    "name": "bcoe"
+  "_nodeVersion": "6.9.5",
+  "_npmOperationalInternal": {
+    "host": "s3://npm-registry-packages",
+    "tmp": "tmp/yargs-8.0.2.tgz_1497310246372_0.5759420988615602"
   },
-  "_npmVersion": "2.9.0",
+  "_npmUser": {
+    "name": "bcoe",
+    "email": "ben@npmjs.com"
+  },
+  "_npmVersion": "4.6.1",
   "_phantomChildren": {},
   "_requested": {
-    "name": "yargs",
-    "raw": "yargs@~3.10.0",
-    "rawSpec": "~3.10.0",
+    "raw": "yargs@^8.0.2",
     "scope": null,
-    "spec": ">=3.10.0 <3.11.0",
+    "escapedName": "yargs",
+    "name": "yargs",
+    "rawSpec": "^8.0.2",
+    "spec": ">=8.0.2 <9.0.0",
     "type": "range"
   },
   "_requiredBy": [
-    "/uglify-js"
+    "/react-native-vector-icons"
   ],
-  "_resolved": "https://registry.npmjs.org/yargs/-/yargs-3.10.0.tgz",
-  "_shasum": "f7ee7bd857dd7c1d2d38c0e74efbd681d1431fd1",
+  "_resolved": "https://registry.npmjs.org/yargs/-/yargs-8.0.2.tgz",
+  "_shasum": "6299a9055b1cefc969ff7e79c1d918dceb22c360",
   "_shrinkwrap": null,
-  "_spec": "yargs@~3.10.0",
-  "_where": "/home/gil/Application/nodejs/sbms/node_modules/uglify-js",
-  "author": {
-    "email": "Alex.Ford@CodeTunnel.com",
-    "name": "Alex Ford",
-    "url": "http://CodeTunnel.com"
-  },
+  "_spec": "yargs@^8.0.2",
+  "_where": "C:\\Android_react\\PPSI-11\\node_modules\\react-native-vector-icons",
   "bugs": {
-    "url": "https://github.com/bcoe/yargs/issues"
+    "url": "https://github.com/yargs/yargs/issues"
   },
-  "contributors": [
-    {
-      "name": "Benjamin Coe",
-      "email": "ben@npmjs.com",
-      "url": "https://github.com/bcoe"
-    },
-    {
-      "name": "Chris Needham",
-      "email": "chris@chrisneedham.com",
-      "url": "http://chrisneedham.com"
-    },
-    {
-      "name": "James Nylen",
-      "email": "jnylen@gmail.com",
-      "url": "https://github.com/nylen"
-    },
-    {
-      "name": "Benjamin Horsleben",
-      "url": "https://github.com/fizker"
-    },
-    {
-      "name": "Lin Clark",
-      "url": "https://github.com/linclark"
-    }
-  ],
   "dependencies": {
-    "camelcase": "^1.0.2",
-    "cliui": "^2.1.0",
-    "decamelize": "^1.0.0",
-    "window-size": "0.1.0"
+    "camelcase": "^4.1.0",
+    "cliui": "^3.2.0",
+    "decamelize": "^1.1.1",
+    "get-caller-file": "^1.0.1",
+    "os-locale": "^2.0.0",
+    "read-pkg-up": "^2.0.0",
+    "require-directory": "^2.1.1",
+    "require-main-filename": "^1.0.1",
+    "set-blocking": "^2.0.0",
+    "string-width": "^2.0.0",
+    "which-module": "^2.0.0",
+    "y18n": "^3.2.1",
+    "yargs-parser": "^7.0.0"
   },
-  "description": "Light-weight option parsing with an argv hash. No optstrings attached.",
+  "description": "yargs the modern, pirate-themed, successor to optimist.",
   "devDependencies": {
-    "chai": "^2.2.0",
-    "coveralls": "^2.11.2",
+    "chai": "^3.4.1",
+    "chalk": "^1.1.3",
+    "coveralls": "^2.11.11",
+    "cpr": "^2.0.0",
+    "cross-spawn": "^5.0.1",
+    "es6-promise": "^4.0.2",
     "hashish": "0.0.4",
-    "mocha": "^2.2.1",
-    "nyc": "^2.2.1",
-    "standard": "^3.11.1"
+    "mocha": "^3.0.1",
+    "nyc": "^10.3.0",
+    "rimraf": "^2.5.0",
+    "standard": "^8.6.0",
+    "standard-version": "^4.2.0",
+    "which": "^1.2.9",
+    "yargs-test-extends": "^1.0.1"
   },
   "directories": {},
   "dist": {
-    "shasum": "f7ee7bd857dd7c1d2d38c0e74efbd681d1431fd1",
-    "tarball": "https://registry.npmjs.org/yargs/-/yargs-3.10.0.tgz"
+    "shasum": "6299a9055b1cefc969ff7e79c1d918dceb22c360",
+    "tarball": "https://registry.npmjs.org/yargs/-/yargs-8.0.2.tgz"
   },
   "engine": {
-    "node": ">=0.4"
+    "node": ">=0.10"
   },
   "files": [
-    "LICENSE",
-    "completion.sh.hbs",
     "index.js",
-    "lib"
+    "yargs.js",
+    "lib",
+    "locales",
+    "completion.sh.hbs",
+    "LICENSE"
   ],
-  "gitHead": "491e6b10e46485a31504e6a1ef21450dff425030",
-  "homepage": "https://github.com/bcoe/yargs#readme",
+  "gitHead": "30a18720759163ec1a5f6ca54ecb9e089da8b5a0",
+  "homepage": "http://yargs.js.org/",
   "keywords": [
-    "args",
     "argument",
-    "cli",
-    "command",
+    "args",
     "option",
     "parser",
-    "parsing"
+    "parsing",
+    "cli",
+    "command"
   ],
   "license": "MIT",
   "main": "./index.js",
   "maintainers": [
     {
+      "name": "bcoe",
+      "email": "ben@npmjs.com"
+    },
+    {
       "name": "chevex",
       "email": "alex.ford@codetunnel.com"
     },
     {
-      "name": "bcoe",
-      "email": "bencoe@gmail.com"
-    },
-    {
-      "name": "nylen",
-      "email": "jnylen@gmail.com"
+      "name": "nexdrew",
+      "email": "andrew@npmjs.com"
     }
   ],
   "name": "yargs",
   "optionalDependencies": {},
-  "readme": "ERROR: No README data found!",
+  "readme": "# Yargs\n\n[![Build Status][travis-image]][travis-url]\n[![Coverage Status][coveralls-image]][coveralls-url]\n[![NPM version][npm-image]][npm-url]\n[![Windows Tests][windows-image]][windows-url]\n[![js-standard-style][standard-image]][standard-url]\n[![Conventional Commits][conventional-commits-image]][conventional-commits-url]\n[![Gitter][gitter-image]][gitter-url]\n\n> Yargs be a node.js library fer hearties tryin' ter parse optstrings.\n\n<img width=\"250\" src=\"/yargs-logo.png\">\n\nYargs helps you build interactive command line tools, by parsing arguments and generating an elegant user interface. It gives you:\n\n* commands and (grouped) options (`my-program.js serve --port=5000`).\n* a dynamically generated help menu based on your arguments.\n\n> <img width=\"400\" src=\"/screen.png\">\n\n* bash-completion shortcuts for commands and options.\n* and [tons more](/docs/api.md).\n\n## Installation\n\n```bash\nnpm i yargs --save\n```\n\n## Simple Example\n\n````javascript\n#!/usr/bin/env node\nconst argv = require('yargs').argv\n\nif (argv.ships > 3 && argv.distance < 53.5) {\n  console.log('Plunder more riffiwobbles!')\n} else {\n  console.log('Retreat from the xupptumblers!')\n}\n````\n\n```bash\n$ ./plunder.js --ships=4 --distance=22\nPlunder more riffiwobbles!\n\n$ ./plunder.js --ships 12 --distance 98.7\nRetreat from the xupptumblers!\n```\n\n## Complex Example\n\n```js\n#!/usr/bin/env node\nconst yargs = require('yargs') // eslint-disable-line\n  .command('serve', 'start the server', (yargs) => {\n    yargs.option('port', {\n      describe: 'port to bind on',\n      default: 5000\n    })    \n  }, (argv) => {\n    if (argv.verbose) console.info(`start server on :${argv.port}`)\n    serve(argv.port)\n  })\n  .option('verbose', {\n    alias: 'v',\n    default: false\n  })\n  .help()\n  .argv\n```\n\n## Table of Contents\n\n* [Yargs' API](/docs/api.md)\n* [Examples](/docs/examples.md)\n* [Parsing Tricks](/docs/tricks.md)\n  * [Stop the Parser](/docs/tricks.md#stop)\n  * [Negating Boolean Arguments](/docs/tricks.md#negate)\n  * [Numbers](/docs/tricks.md#numbers)\n  * [Arrays](/docs/tricks.md#arrays)\n  * [Objects](/docs/tricks.md#objects)\n* [Advanced Topics](/docs/advanced.md)\n  * [Composing Your App Using Commands](/docs/advanced.md#commands)\n  * [Building Configurable CLI Apps](/docs/advanced.md#configuration)\n  * [Customizing Yargs' Parser](/docs/advanced.md#customizing)\n* [Contributing](/contributing.md)\n\n[travis-url]: https://travis-ci.org/yargs/yargs\n[travis-image]: https://img.shields.io/travis/yargs/yargs/master.svg\n[coveralls-url]: https://coveralls.io/github/yargs/yargs\n[coveralls-image]: https://img.shields.io/coveralls/yargs/yargs.svg\n[npm-url]: https://www.npmjs.com/package/yargs\n[npm-image]: https://img.shields.io/npm/v/yargs.svg\n[windows-url]: https://ci.appveyor.com/project/bcoe/yargs-ljwvf\n[windows-image]: https://img.shields.io/appveyor/ci/bcoe/yargs-ljwvf/master.svg?label=Windows%20Tests\n[standard-image]: https://img.shields.io/badge/code%20style-standard-brightgreen.svg\n[standard-url]: http://standardjs.com/\n[conventional-commits-image]: https://img.shields.io/badge/Conventional%20Commits-1.0.0-yellow.svg\n[conventional-commits-url]: https://conventionalcommits.org/\n[gitter-image]: https://img.shields.io/gitter/room/nwjs/nw.js.svg?maxAge=2592000\n[gitter-url]: https://gitter.im/yargs/Lobby?utm_source=share-link&utm_medium=link&utm_campaign=share-link\n",
+  "readmeFilename": "README.md",
   "repository": {
     "type": "git",
-    "url": "git+ssh://git@github.com/bcoe/yargs.git"
+    "url": "git+ssh://git@github.com/yargs/yargs.git"
   },
   "scripts": {
     "coverage": "nyc report --reporter=text-lcov | coveralls",
-    "test": "standard && nyc mocha --check-leaks && nyc report"
+    "pretest": "standard",
+    "release": "standard-version",
+    "test": "nyc --cache mocha --require ./test/before.js --timeout=8000 --check-leaks"
   },
   "standard": {
-    "globals": [
-      "it"
-    ],
     "ignore": [
       "**/example/**"
     ]
   },
-  "version": "3.10.0"
+  "version": "8.0.2"
 }
